Fix for relocation of proto api
